TD2: Added docstrings to main, broken out json methods, reordered main methods.
