+++ src/main.py
@@ -2,12 +2,12 @@ import requests
 import pandas as pd
 
 import configparser
-import logging
 import json
 import os
 import datetime as dt
 
-from src.config_tools import *
+from src.utilities.config_tools import *
+import src.utilities.json_tools as jt
 
 
 def raise_request(url: str, headers: dict = None) -> requests.Response:
@@ -18,7 +18,7 @@ def raise_request(url: str, headers: dict = None) -> requests.Response:
     url : str
         Address to access
     headers : dict
-        Header values to pass into requests.get to alter the response from
+        Header values to pass into "requests.get" to alter the response from
         the url
         The default is None
 
@@ -40,62 +40,128 @@ def raise_request(url: str, headers: dict = None) -> requests.Response:
     return r
 
 
-def jsonify_response(r: requests.Response, encoding: str = 'utf-8-sig'):
+def scrape_mp_data(config: configparser.ConfigParser):
     """
-    Takes a requests.Response object, encodes to utf-8-sig by default,
-    and returns a dictionary containing the data in a json format
+    Fetches a copy of the raw MP data, from gov.uk.
 
     Parameters
     ----------
-    r : requests.Response
-        Response object from requests module, containing json data.
-    encoding : str
-        Specific encoding for received data.
-        The default is 'utf-8-sig'.
+    config : configparser.ConfigParser
+        Contents of config file.
+    Returns
+    -------
+    pd.DataFrame
+        Full MP data, from api.
+    """
+    logging.debug('Getting data from api')
+
+    json_path = config['PARSER']['mp_json_path']
+    url = config['PARSER']['members_url']
+
+    r = raise_request(url, headers=cfg_get_dict(config, 'PARSER', 'headers'))
+
+    data, r = jt.jsonify_response(r)
+    jt.write_to_json(data, json_path)
+
+    full_member_data = jt.json_to_pd(data['Members']['Member'],
+                                     cols={'@Member_Id': 'int64'})
+
+    return full_member_data
+
+
+def read_local_mp_data(config: configparser.ConfigParser):
+    """
+    Fetches a copy of the raw MP data, from local copy created previously.
+
+    Parameters
+    ----------
+    config : configparser.ConfigParser
+        Contents of config file.
 
     Returns
     -------
-    tuple(data, r)
-        Data is the contents converted to a json format - stored in a
-        dictionary several levels deep.
-        r is the requests.Response value (encoded)
+    pd.DataFrame
+        Full MP data, from local.
     """
-    logging.debug('Converting response data to json style data as a dict')
-    r.encoding = encoding
-    data = r.json()
+    logging.debug('Getting data from file')
 
-    return data, r
+    json_path = config['PARSER']['mp_json_path']
+
+    def data_exists(path): return os.path.exists(path)
+    assert data_exists(json_path), \
+        logging.error('File {0} missing. Rerun in scrape mode by '
+                      'setting scrape_mode in config'.format(json_path))
+
+    with open(json_path) as f:
+        data = json.load(f)
+
+    full_member_data = jt.json_to_pd(data['Members']['Member'],
+                                     cols={'@Member_Id': 'int64'})
+
+    return full_member_data
 
 
-def write_to_json(data: dict,
-                  file_name: str,
-                  write_mode: str = 'w'):
+def get_mp_data(config: configparser.ConfigParser):
     """
-    Takes Response, encodes, pulls out data as json and writes to file
+    Fetches a copy of the raw MP data, either from gov.uk or from local copy
+    created previously.
+
     Parameters
     ----------
-    data: dict
-        Response data converted to a dict storing json style data
-    file_name: str
-        File name to write out data as.
-    write_mode: str
-        Opens file_name in specific mode.
-        The default is 'w'
+    config : configparser.ConfigParser
+        Contents of config file.
+
+    Returns
+    -------
+    pd.DataFrame
+        Full MP data, from either api or local depending on config settings.
     """
+    scrape_mode = config['PARSER']['scrape_mode']
 
-    logging.info('Dumping to {0}'.format(file_name))
-
-    with open(file_name, write_mode) as f:
-        json.dump(data, f, indent=2)
-
-    exist_status = os.path.exists(file_name)
-
-    if exist_status:
-        time = dt.datetime.utcfromtimestamp(os.path.getmtime(file_name))
-        logging.debug('Dump complete. Status unknown: File exists, '
-                      'last modified: {0}.'.format(time))
+    if scrape_mode == 'True':
+        full_member_data = scrape_mp_data(config)
     else:
-        logging.warning('Dump failed. Status unknown - File does not exist.')
+        full_member_data = read_local_mp_data(config)
+
+    logging.debug('Found full mp data of shape: {0}'.format(
+        full_member_data.shape))
+
+    return full_member_data
+
+
+def clean_mp_data(config: configparser.ConfigParser,
+                  full_member_data: pd.DataFrame):
+    """
+    Clean raw MP data:
+    - Keep only needed columns
+    - Create a clean name column without titles or honourifics
+
+    Parameters
+    ----------
+    config : configparser.ConfigParser
+        Contents of config file.
+    full_member_data : pd.DataFrame
+        Unclean MP data, with many superfluous columns.
+
+    Returns
+    -------
+    pd.DataFrame
+        Cleaned MP data.
+
+    """
+    # Keep only desired columns.
+    mp_cols = cfg_get_list(config, 'PARSER', 'mp_cols')
+    core_mp_data = full_member_data[mp_cols]
+
+    # Create clean name column - no titles or honourifics.
+    titles = cfg_get_list(config, 'PARSER', 'titles')
+    core_mp_data['clearName'] = core_mp_data['DisplayAs'].str.replace(
+        '|'.join(titles), '').str.strip()
+
+    logging.debug('Cut down to core shape: {0} and core columns: {1}'.format(
+        core_mp_data.shape, core_mp_data.columns))
+
+    return core_mp_data
 
 
 def get_characteristics():
@@ -174,7 +240,7 @@ def add_characteristics(config: configparser.ConfigParser,
 
     Parameters
     ----------
-    config: configparser.ConfigParser
+    config : configparser.ConfigParser
         Contents of config file
     member_data : pd.DataFrame
         MP member data from either scraped gov website, or local copy.
@@ -224,86 +290,6 @@ def add_characteristics(config: configparser.ConfigParser,
         report_missing_characteristics(config, missing_characteristics)
 
     return matched_members
-
-
-def json_to_pd(json_data: list, cols: dict, max_level=1):
-
-    pd_data = pd.json_normalize(json_data, max_level=1)
-
-    for col, col_type in cols.items():
-        pd_data[col] = pd_data[col].astype(col_type)
-
-    return pd_data
-
-
-def scrape_mp_data(config: configparser.ConfigParser):
-    logging.debug('Getting data from api')
-
-    json_path = config['PARSER']['mp_json_path']
-    url = config['PARSER']['members_url']
-
-    r = raise_request(url, headers=cfg_get_dict(config, 'PARSER', 'headers'))
-
-    data, r = jsonify_response(r)
-    write_to_json(data, json_path)
-
-    full_member_data = json_to_pd(data['Members']['Member'],
-                                  cols={'@Member_Id': 'int64'})
-
-    return full_member_data
-
-
-def read_local_mp_data(config: configparser.ConfigParser):
-    logging.debug('Getting data from file')
-
-    json_path = config['PARSER']['mp_json_path']
-
-    def data_exists(path): return os.path.exists(path)
-
-    assert data_exists(json_path), \
-        logging.error('File {0} missing. Rerun in scrape mode by '
-                      'setting scrape_mode in config'.format(json_path))
-
-    with open(json_path) as f:
-        data = json.load(f)
-
-    full_member_data = json_to_pd(data['Members']['Member'],
-                                  cols={'@Member_Id': 'int64'})
-
-    return full_member_data
-
-
-def get_mp_data(config: configparser.ConfigParser):
-
-    scrape_mode = config['PARSER']['scrape_mode']
-
-    if scrape_mode == 'True':
-        full_member_data = scrape_mp_data(config)
-    else:
-        full_member_data = read_local_mp_data(config)
-
-    logging.debug('Found full mp data of shape: {0}'.format(
-        full_member_data.shape))
-
-    return full_member_data
-
-
-def clean_mp_data(config: configparser.ConfigParser,
-                  full_member_data: pd.DataFrame):
-
-    # Keep only desired columns.
-    mp_cols = cfg_get_list(config, 'PARSER', 'mp_cols')
-    core_mp_data = full_member_data[mp_cols]
-
-    # Create clean name column - no titles or honourifics.
-    titles = cfg_get_list(config, 'PARSER', 'titles')
-    core_mp_data['clearName'] = core_mp_data['DisplayAs'].str.replace(
-        '|'.join(titles), '').str.strip()
-
-    logging.debug('Cut down to core shape: {0} and core columns: {1}'.format(
-        core_mp_data.shape, core_mp_data.columns))
-
-    return core_mp_data
 
 
 def main(config: configparser.ConfigParser):
